Move read related classes to separate namespace
