+++ spec/Broadway/SourceFactory/Read/ElasticSearch/ReadRepositorySourceFactorySpec.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace spec\NullDev\Skeleton\Broadway\SourceFactory;
+namespace spec\NullDev\Skeleton\Broadway\SourceFactory\Read\ElasticSearch;
 
-use NullDev\Skeleton\Broadway\SourceFactory\ReadRepositorySourceFactory;
+use NullDev\Skeleton\Broadway\SourceFactory\Read\ElasticSearch\ReadRepositorySourceFactory;
 use NullDev\Skeleton\Definition\PHP\DefinitionFactory;
 use NullDev\Skeleton\Definition\PHP\Types\ClassType;
 use NullDev\Skeleton\Source\ClassSourceFactory;
