+++ tests/Broadway/CodeGenerator/BaseCodeGeneratorTest.php
@@ -8,9 +8,9 @@ use NullDev\Skeleton\Broadway\SourceFactory\CommandSourceFactory;
 use NullDev\Skeleton\Broadway\SourceFactory\EventSourcedAggregateRootSourceFactory;
 use NullDev\Skeleton\Broadway\SourceFactory\EventSourceFactory;
 use NullDev\Skeleton\Broadway\SourceFactory\EventSourcingRepositorySourceFactory;
-use NullDev\Skeleton\Broadway\SourceFactory\ReadEntitySourceFactory;
-use NullDev\Skeleton\Broadway\SourceFactory\ReadProjectorSourceFactory;
-use NullDev\Skeleton\Broadway\SourceFactory\ReadRepositorySourceFactory;
+use NullDev\Skeleton\Broadway\SourceFactory\Read\ElasticSearch\ReadEntitySourceFactory;
+use NullDev\Skeleton\Broadway\SourceFactory\Read\ElasticSearch\ReadProjectorSourceFactory;
+use NullDev\Skeleton\Broadway\SourceFactory\Read\ElasticSearch\ReadRepositorySourceFactory;
 use NullDev\Skeleton\Definition\PHP\DefinitionFactory;
 use NullDev\Skeleton\Definition\PHP\Parameter;
 use NullDev\Skeleton\Definition\PHP\Types\ClassType;
